requirements.txt
langchain
pandas
numpy
streamlit
openai
python-dotenv